Refactor to push the code into methods of PairsData

+++ src/syzygy_probe.hpp
@@ -18,7 +18,7 @@ struct TbId{
     
     constexpr bool symmetric() const { return stronger == weaker; }
     constexpr bool has_pawns() const { return ((stronger & 7) == 1) or ((weaker & 7) == 1); }
-    int num() const;  // Total number of pieces, including kings
+    size_t num() const;  // Total number of pieces, including kings
     bool enc_type_2() const;  // Use the K2 piece encoding
 
     std::tuple<int, int> pawn_counts() const;
